+++ R/plot.R
@@ -87,7 +87,7 @@ kw_dflt = function(x, key.pos) {
 plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes = FALSE,
 		downsample = TRUE, nbreaks = 11, breaks = "quantile", col = grey(1:(nbreaks-1)/nbreaks),
 		key.pos = get_key_pos(x, ...), key.width = kw_dflt(x, key.pos), key.length = 0.618, 
-		key.lab = main, reset = TRUE, box_col = NA, center_time = FALSE, hook = NULL, 
+		key.lab = ifelse(length(main) == 1, main, ""), reset = TRUE, box_col = NA, center_time = FALSE, hook = NULL, 
 		mfrow = NULL, compact = TRUE) {
 
 	if (!missing(y))
@@ -156,10 +156,11 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 					get_breaks(x, breaks, nbreaks, dots$logz)
 		if (!inherits(breaks, c("POSIXt", "Date")))
 			breaks = as.numeric(breaks)
-		if (length(breaks) > 2)
+		# https://github.com/r-spatial/stars/issues/728:
+		if (packageVersion("sf") < "1.0.20" && length(breaks) > 2)
 			breaks = unique(breaks)
 		nbreaks = length(breaks) # might be shorter than originally intended!
-		if (breaks.missing && nbreaks <= 2) # unlucky default!
+		if (breaks.missing && length(unique(breaks)) <= 2) # unlucky default!
 			warning('breaks="quantile" leads to a single class; maybe try breaks="equal" instead?')
 	}