+++ R/stars.R
@@ -269,20 +269,20 @@ xy_from_colrow = function(x, geotransform) {
 		x %*% matrix(geotransform[c(2, 3, 5, 6)], nrow = 2, ncol = 2)
 }
 
-colrow_from_xy = function(x, obj, NA_outside = FALSE) {
+colrow_from_xy = function(x, obj, NA_outside = FALSE, flip = FALSE) {
 	if (inherits(obj, "stars"))
 		obj = st_dimensions(obj)
 	xy = attr(obj, "raster")$dimensions
 	if (inherits(obj, "dimensions"))
 		gt = st_geotransform(obj)
 
-	if (isTRUE(st_is_longlat(st_crs(obj)))) {
+	if (flip && isTRUE(st_is_longlat(st_crs(obj)))) {
 		bb = st_bbox(obj)
 # see https://github.com/r-spatial/stars/issues/519 where this is problematic;
 # not sure whether this introduces new problems.
 #		sign = ifelse(x[,1] < bb["xmin"], 1., ifelse(x[,1] > bb["xmax"], -1., 0.))
 #		x[,1] = x[,1] + sign * 360.
-		# one more try: https://github.com/r-spatial/stars/issues/563
+		## one more try: https://github.com/r-spatial/stars/issues/563
 		ix = x[,1] > bb["xmax"] & !is.na(x[,1])
 		x[ix,1] = x[ix,1] - 360.
 		ix = x[,1] < bb["xmin"] & !is.na(x[,1])
@@ -474,11 +474,11 @@ st_coordinates.dimensions = function(x, ...) {
 #' @param add_coordinates logical; if `TRUE`, columns with dimension values preceed the array values, 
 #' otherwise they are omitted
 as.data.frame.stars = function(x, ..., add_max = FALSE, center = NA, add_coordinates = TRUE) {
+	df = setNames(as.data.frame(lapply(x, function(y) structure(y, dim = NULL))), names(x))
 	if (add_coordinates)
-		data.frame(st_coordinates(x, add_max = add_max, center = center, ...), 
-			lapply(x, function(y) structure(y, dim = NULL)))
+		cbind(st_coordinates(x, add_max = add_max, center = center, ...), df)
 	else
-		as.data.frame(lapply(x, function(y) structure(y, dim = NULL)))
+		df
 }
 
 add_units = function(x) {
@@ -905,13 +905,20 @@ merge.stars = function(x, y, ..., name = "attributes") {
 }
 
 sort_out_along = function(ret) { 
-	d1 = st_dimensions(ret[[1]])
-	d2 = st_dimensions(ret[[2]])
-	if ("time" %in% names(d1) && (isTRUE(d1$time$offset != d2$time$offset) || 
-			!any(d1$time$values %in% d2$time$values)))
-		"time"
-	else
-		NA_integer_
+	# https://github.com/r-spatial/stars/issues/703 :
+	# 1. check that time is a dimension name in all objects 
+	l = lapply(ret, st_dimensions)
+	if (!all(sapply(l, function(x) "time" %in% names(x))))
+		return(NA_integer_)
+	# 2. check that time values do not overlap
+	lv = lapply(l, st_get_dimension_values, "time")
+	for (i in seq_along(lv)) {
+		if (!inherits(lv[[i]], c("POSIXt", "Date", "PCICt")))
+			return(NA_integer_)
+		if (i < length(lv) && max(lv[[i]]) >= min(lv[[i+1]])) # no sequence
+			return(NA_integer_)
+	}
+	return("time")
 }
 
 
@@ -967,8 +974,12 @@ st_redimension.stars = function(x, new_dims = st_dimensions(x),
 			dims = create_dimensions(c(d, new_dim = list(new_dim)), attr(d, "raster"))
 			if (length(names(along)) == 1)
 				names(dims)[names(dims) == "new_dim"] = names(along)
-			ret = structure(do.call(c, x), dim = dim(dims))
-			st_stars(setNames(list(ret), paste(names(x), collapse = ".")), dimensions = dims)
+			ns = names(x)
+			x = if (inherits(x[[1]], c("factor", "Date", "POSIXt")))
+					structure(do.call(c, x), dim = dim(dims)) # don't do this on large data sets!!
+				else
+					do.call(abind, append(unclass(x), list(along = length(dims))))
+			st_stars(setNames(list(x), paste(ns, collapse = ".")), dimensions = dims)
 		}
 	}
 }