+++ NEWS.md
@@ -1,4 +1,36 @@
+# version 0.6-9
+
+* `slice.stars_proxy()` works (again) for single-file proxy objects; #751, #527
+
+* `[.stars()` with a `character` selector selects on names of a dimension if it has names; #747
+
+* handle "intervals" dimension in aggregate.stars() and `st_as_sf.stars()`; #745
+
+* fix `c.stars()` for the case where some of the objects have a single slice on the `along=` dimension; #743
+
+* `st_redimension()` (and by that, `merge.stars()`) use `abind()` for numeric variables rather than `c()`, reducing memory overhead.
+
+* fix problem with reading HDF4 files; #741 by Alexys Rodriguez
+
+* use, by default, `normalizePath()` on `filename` in `read_mdim()`;  #735
+
+* `st_extract()` accepts empty points as target; #734
+
+* `st_sfc2xy()` passes `...` on to `st_as_stars()`; #733 
+
+# version 0.6-8
+
+* address `/.difftime` issue new in R-devel rev 87670
+
+* `c.stars()` is more strict when combining time sequences; #703
+
+* fix plotting when breaks contain duplicates; #728
+
+* fix `st_as_stars.im()`; #727 and #648, thanks to Barry Rowlingson
+
 # version 0.6-7
+
+* `st_extract()` fix if points coincide with boundary grid cell centers and bilinear interpolation is used; #720
 
 * `st_extract()` if used with GDAL 3.10.0 uses InterpolateAtPoints, allowing for cubic and cubicspline interpolators (requiring sf >= 1.0-19).